Move unit_bypass_validation tests to unit/
